Hasta donde llegué, me falta guardar los detalles de la factura. Front inicializado, se debe realizar el enrutamiento de forma correcta, fallo en renderización

+++ src/main/java/com/example/demo/repository/ClienteRepository.java
@@ -7,6 +7,6 @@ import org.springframework.transaction.annotation.Transactional;
 //import java.util.List;
 
 @Transactional
-public interface ClienteRepository extends JpaRepository<Cliente,Long> {
-    //List<Cliente> findByObjectId(Long idCliente);
+public interface ClienteRepository extends JpaRepository<Cliente,Integer> {
+    //List<Cliente> findByTipoidentificacion(Integer tIden);
 }
